@@ -1,7 +1,8 @@
 {
+  "type": "module",
   "name": "learning-platform",
   "version": "1.0.0",
-  "description": "Learning Platform - Production Build",
+  "description": "Learning Platform - Production Ready",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -10,7 +11,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "^14.0.0",
+    "next": "^14.2.30",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.0",
