@@ -1,29 +1,66 @@
+import jwt from 'jsonwebtoken';
 import bcrypt from 'bcryptjs';
 import { query } from './db.js';
 
-const JWT_SECRET = process.env.JWT_SECRET || 'fallback-secret-key';
+const JWT_SECRET = process.env.JWT_SECRET || 'your-fallback-secret-key';
 
-// توليد معرف مستخدم فريد
-export async function generateUserId() {
-  try {
-    const result = await query('SELECT MAX(user_id) as max_id FROM users');
-    return result[0]?.max_id ? result[0].max_id + 1 : 1000;
-  } catch (error) {
-    console.error('Error generating user ID:', error);
-    return 1000;
-  }
+// إنشاء توكن JWT
+export function generateToken(user) {
+  return jwt.sign(
+    { 
+      user_id: user.user_id,
+      phone: user.phone,
+      id: user.id 
+    },
+    JWT_SECRET,
+    { expiresIn: '30d' }
+  );
 }
 
-// توليد توكن
-export function generateToken(userId, phone) {
-  const randomString = Math.random().toString(36).substring(2, 15) + 
-                      Math.random().toString(36).substring(2, 15);
-  return `${userId}|${randomString}`;
+// التحقق من صحة التوكن
+export async function verifyToken(authHeader) {
+  try {
+    if (!authHeader) {
+      throw new Error('No authorization header provided');
+    }
+
+    const token = authHeader.startsWith('Bearer ') 
+      ? authHeader.slice(7) 
+      : authHeader;
+
+    if (!token) {
+      throw new Error('No token provided');
+    }
+
+    const decoded = jwt.verify(token, JWT_SECRET);
+    
+    const users = await query(
+      'SELECT * FROM users WHERE user_id = ? AND active = TRUE',
+      [decoded.user_id]
+    );
+
+    if (users.length === 0) {
+      throw new Error('User not found or inactive');
+    }
+
+    return {
+      ...decoded,
+      id: users[0].id,
+      user_id: users[0].user_id,
+      phone: users[0].phone,
+      created_at: users[0].created_at
+    };
+
+  } catch (error) {
+    console.error('Token verification error:', error.message);
+    throw new Error('Invalid or expired token');
+  }
 }
 
 // تشفير كلمة المرور
 export async function hashPassword(password) {
-  return await bcrypt.hash(password, 10);
+  const saltRounds = 12;
+  return await bcrypt.hash(password, saltRounds);
 }
 
 // مقارنة كلمة المرور
@@ -31,28 +68,7 @@ export async function comparePassword(password, hashedPassword) {
   return await bcrypt.compare(password, hashedPassword);
 }
 
-// التحقق من رقم الهاتف العراقي
-export function validateIraqiPhone(phone) {
-  const phoneRegex = /^07[0-9]{9}$/;
-  return phoneRegex.test(phone);
-}
-
-// التحقق من التوكن
-export async function verifyToken(token) {
-  if (!token || !token.startsWith('Bearer ')) {
-    return null;
-  }
-  
-  try {
-    const actualToken = token.substring(7);
-    const result = await query(
-      'SELECT * FROM users WHERE token = ? AND active = TRUE', 
-      [actualToken]
-    );
-    
-    return result[0] || null;
-  } catch (error) {
-    console.error('Token verification error:', error);
-    return null;
-  }
+// إنشاء معرف مستخدم عشوائي
+export function generateUserId() {
+  return Math.floor(100000 + Math.random() * 900000).toString();
 }
