@@ -1,9 +1,13 @@
 import { killSleepingConnections } from './db.js';
 
-// تنظيف الاتصالات كل 30 ثانية
 let cleanupInterval = null;
 
 export function startConnectionCleanup() {
+  // فقط في بيئة الإنتاج
+  if (process.env.NODE_ENV !== 'production') {
+    return;
+  }
+  
   if (cleanupInterval) {
     clearInterval(cleanupInterval);
   }
@@ -19,7 +23,7 @@ export function startConnectionCleanup() {
     } catch (error) {
       console.error('[CLEANUP] Cleanup failed:', error.message);
     }
-  }, 30000); // كل 30 ثانية
+  }, 60000); // كل دقيقة
   
   console.log('[CLEANUP] Connection cleanup started');
 }
@@ -33,14 +37,5 @@ export function stopConnectionCleanup() {
 }
 
 // تنظيف عند إغلاق العملية
-process.on('SIGINT', () => {
-  console.log('[CLEANUP] Received SIGINT, cleaning up...');
-  stopConnectionCleanup();
-  process.exit(0);
-});
-
-process.on('SIGTERM', () => {
-  console.log('[CLEANUP] Received SIGTERM, cleaning up...');
-  stopConnectionCleanup();
-  process.exit(0);
-});
+process.on('SIGINT', stopConnectionCleanup);
+process.on('SIGTERM', stopConnectionCleanup);
