@@ -1,48 +1,6 @@
 import mysql from 'mysql2/promise';
 
-// إنشاء اتصال واحد مشترك بدلاً من pool
-let connection = null;
-
-// دالة للحصول على الاتصال
-async function getConnection() {
-  if (!connection || connection.connection._socket.destroyed) {
-    console.log('[DB] Creating new database connection...');
-    
-    connection = await mysql.createConnection({
-      host: process.env.DB_HOST || 'localhost',
-      user: process.env.DB_USER || 'root',
-      password: process.env.DB_PASSWORD || '',
-      database: process.env.DB_NAME || 'learning_platform',
-      port: process.env.DB_PORT || 3306,
-      
-      // إعدادات لمنع الاتصالات النائمة
-      acquireTimeout: 10000,     // 10 ثواني فقط للحصول على اتصال
-      timeout: 10000,            // 10 ثواني فقط للاستعلام
-      reconnect: false,          // عدم إعادة الاتصال تلقائياً
-      multipleStatements: false,
-      charset: 'utf8mb4',
-      
-      // إعدادات مهمة لقتل الاتصالات النائمة
-      typeCast: function (field, next) {
-        return next();
-      }
-    });
-    
-    // إضافة مستمعين للأحداث
-    connection.on('error', (err) => {
-      console.error('[DB] Connection error:', err);
-      if (err.code === 'PROTOCOL_CONNECTION_LOST') {
-        connection = null;
-      }
-    });
-    
-    console.log('[DB] Database connection established');
-  }
-  
-  return connection;
-}
-
-// دالة الاستعلام مع إغلاق فوري
+// دالة الاستعلام مع إغلاق فوري للاتصالات
 export async function query(sql, params = []) {
   const startTime = Date.now();
   let conn = null;
@@ -54,9 +12,12 @@ export async function query(sql, params = []) {
       user: process.env.DB_USER || 'root',
       password: process.env.DB_PASSWORD || '',
       database: process.env.DB_NAME || 'learning_platform',
-      port: process.env.DB_PORT || 3306,
-      timeout: 5000,             // 5 ثواني فقط
-      acquireTimeout: 5000,      // 5 ثواني فقط
+      port: parseInt(process.env.DB_PORT) || 3306,
+      timeout: 5000,
+      acquireTimeout: 5000,
+      charset: 'utf8mb4',
+      reconnect: false,
+      multipleStatements: false
     });
     
     console.log(`[DB] New connection created for query`);
@@ -91,16 +52,14 @@ export async function killSleepingConnections() {
   let adminConn = null;
   
   try {
-    // اتصال إداري لقتل الاتصالات
     adminConn = await mysql.createConnection({
       host: process.env.DB_HOST || 'localhost',
       user: process.env.DB_USER || 'root',
       password: process.env.DB_PASSWORD || '',
       database: process.env.DB_NAME || 'learning_platform',
-      port: process.env.DB_PORT || 3306,
+      port: parseInt(process.env.DB_PORT) || 3306,
     });
     
-    // جلب جميع الاتصالات النائمة
     const [processes] = await adminConn.execute(`
       SELECT ID, COMMAND, TIME, STATE 
       FROM INFORMATION_SCHEMA.PROCESSLIST 
@@ -109,11 +68,10 @@ export async function killSleepingConnections() {
     
     console.log(`[DB] Found ${processes.length} sleeping connections to kill`);
     
-    // قتل كل اتصال نائم
     for (const process of processes) {
       try {
         await adminConn.execute(`KILL CONNECTION ?`, [process.ID]);
-        console.log(`[DB] Killed sleeping connection ${process.ID} (${process.TIME}s)`);
+        console.log(`[DB] Killed sleeping connection ${process.ID}`);
       } catch (killError) {
         console.log(`[DB] Could not kill connection ${process.ID}:`, killError.message);
       }
@@ -142,7 +100,7 @@ export async function getConnectionStatus() {
       user: process.env.DB_USER || 'root',
       password: process.env.DB_PASSWORD || '',
       database: process.env.DB_NAME || 'learning_platform',
-      port: process.env.DB_PORT || 3306,
+      port: parseInt(process.env.DB_PORT) || 3306,
     });
     
     const [processes] = await conn.execute(`
